Move router usage out of canvasOperation and workflowHelpers
