+++ packages/frontend/editor-ui/src/components/ParameterInputList.vue
@@ -39,7 +39,6 @@ import { computedWithControl } from '@vueuse/core';
 import get from 'lodash/get';
 import set from 'lodash/set';
 import { N8nIcon, N8nIconButton, N8nInputLabel, N8nNotice, N8nText } from '@n8n/design-system';
-import { useRouter } from 'vue-router';
 import { storeToRefs } from 'pinia';
 
 const LazyFixedCollectionParameter = defineAsyncComponent(
@@ -75,8 +74,7 @@ const ndvStore = useNDVStore();
 
 const nodeHelpers = useNodeHelpers();
 const asyncLoadingError = ref(false);
-const router = useRouter();
-const workflowHelpers = useWorkflowHelpers({ router });
+const workflowHelpers = useWorkflowHelpers();
 const i18n = useI18n();
 
 const { activeNode } = storeToRefs(ndvStore);
