+++ packages/frontend/editor-ui/src/composables/useNodeDirtiness.test.ts
@@ -15,12 +15,7 @@ import { type FrontendSettings } from '@n8n/api-types';
 import { createTestingPinia } from '@pinia/testing';
 import { NodeConnectionTypes, type IConnections, type IRunData } from 'n8n-workflow';
 import { defineComponent } from 'vue';
-import {
-	createRouter,
-	createWebHistory,
-	useRouter,
-	type RouteLocationNormalizedLoaded,
-} from 'vue-router';
+import { createRouter, createWebHistory, type RouteLocationNormalizedLoaded } from 'vue-router';
 
 describe(useNodeDirtiness, () => {
 	let nodeTypeStore: ReturnType<typeof useNodeTypesStore>;
@@ -42,7 +37,7 @@ describe(useNodeDirtiness, () => {
 				workflowsStore = useWorkflowsStore();
 				settingsStore = useSettingsStore();
 				historyHelper = useHistoryHelper({} as RouteLocationNormalizedLoaded);
-				canvasOperations = useCanvasOperations({ router: useRouter() });
+				canvasOperations = useCanvasOperations();
 				uiStore = useUIStore();
 
 				nodeTypeStore.setNodeTypes(defaultNodeDescriptions);
