+++ packages/frontend/editor-ui/src/plugins/codemirror/completions/utils.ts
@@ -210,7 +210,7 @@ export function autocompletableNodeNames() {
 
 	const activeNodeName = activeNode.name;
 
-	const workflow = useWorkflowHelpers({ router: useRouter() }).getCurrentWorkflow();
+	const workflow = useWorkflowHelpers().getCurrentWorkflow();
 	const nonMainChildren = workflow.getChildNodes(activeNodeName, 'ALL_NON_MAIN');
 
 	// This is a tool node, look for the nearest node with main connections
@@ -222,7 +222,7 @@ export function autocompletableNodeNames() {
 }
 
 export function getPreviousNodes(nodeName: string) {
-	const workflow = useWorkflowHelpers({ router: useRouter() }).getCurrentWorkflow();
+	const workflow = useWorkflowHelpers().getCurrentWorkflow();
 	return workflow
 		.getParentNodesByDepth(nodeName)
 		.map((node) => node.name)
